requirements.txt
mysql-connector-python==8.0.27
numpy==1.21.4
pandas==1.3.4
protobuf==3.19.1
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
